Move all existing endpoints to newer, clearner, split endpoints
