+++ app/api/urls/notifications_urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from api.views.notifications import NotificationsApi
+
+urlpatterns = [
+    path("get/", NotificationsApi.get)
+]
