+++ app/api/urls/sessions_urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from api.views.sessions_api import SessionsApi
+
+urlpatterns = [
+    path("play_start/", SessionsApi.play_start),
+    path("play_save/", SessionsApi.play_save),
+    path("author_verify/", SessionsApi.author_verify),
+]
