+++ app/api/urls/widget_instance_urls.py
@@ -6,5 +6,7 @@ urlpatterns = [
     path("publish_perms_verify/", WidgetInstanceAPI.publish_perms_verify),
     path("save/", WidgetInstanceAPI.save),
     path("update/", WidgetInstanceAPI.update),
-    path("lock/", WidgetInstanceAPI.widget_instance_lock),
+    path("lock/", WidgetInstanceAPI.lock),
+    path("get/", WidgetInstanceAPI.get),
+    path("get_question_set/", WidgetInstanceAPI.get_qset),
 ]
